@@ -1,38 +1,44 @@
 {
-	"name": "evidence-ir",
-	"version": "0.0.1",
+	"name": "seir",
+	"version": "1.1.4",
 	"private": true,
 	"scripts": {
 		"start": "node -r dotenv/config dist",
 		"dev": "vite dev",
 		"build": "vite build",
+		"preview": "vite preview",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --plugin-search-dir . --check . && eslint .",
 		"format": "prettier --plugin-search-dir . --write ."
 	},
 	"devDependencies": {
-		"@sveltejs/kit": "^2.5.25",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
-		"@tsconfig/svelte": "^5.0.4",
+		"@eslint/compat": "^1.2.8",
+		"@eslint/js": "^9.19.0",
+		"@sveltejs/adapter-vercel": "^5.7.0",
+		"@sveltejs/kit": "^2.20.7",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@types/bootstrap": "^5.2.10",
 		"@types/downloadjs": "^1.4.6",
 		"@types/html-to-text": "^9.0.4",
-		"@types/lodash-es": "^4.17.12",
-		"@types/nodemailer": "^6.4.15",
-		"@typescript-eslint/eslint-plugin": "^8.3.0",
-		"@typescript-eslint/parser": "^8.3.0",
-		"@vite-pwa/sveltekit": "^0.6.1",
-		"eslint": "^9.9.1",
-		"eslint-plugin-svelte": "^2.43.0",
-		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.6",
-		"svelte": "^4.2.19",
-		"svelte-check": "^3.8.6",
-		"svelte-preprocess": "^6.0.2",
-		"tslib": "^2.7.0",
-		"typescript": "^5.5.4",
-		"vite": "^5.4.2"
+		"@types/lodash.isstring": "^4.0.9",
+		"@types/lodash.merge": "^4.6.9",
+		"@types/nodemailer": "^6.4.17",
+		"@vite-pwa/sveltekit": "^0.6.8",
+		"autoprefixer": "^10.4.21",
+		"eslint": "^9.25.1",
+		"eslint-config-prettier": "^10.1.2",
+		"eslint-plugin-svelte": "^2.46.1",
+		"globals": "^15.15.0",
+		"prettier": "^3.5.3",
+		"prettier-plugin-svelte": "^3.3.3",
+		"sass": "^1.87.0",
+		"svelte": "^5.28.2",
+		"svelte-check": "^4.1.6",
+		"svelte-preprocess": "^6.0.3",
+		"typescript": "^5.8.3",
+		"typescript-eslint": "^8.31.0",
+		"vite": "^6.3.2"
 	},
 	"type": "module",
 	"dependencies": {
@@ -40,16 +46,17 @@
 		"@sveltejs/adapter-node": "^5.2.2",
 		"dotenv": "^16.4.5",
 		"@popperjs/core": "^2.11.8",
-		"bootstrap": "^5.3.3",
+		"bootstrap": "^5.3.5",
 		"downloadjs": "^1.4.7",
-		"firebase": "^10.13.1",
-		"firebase-admin": "^12.4.0",
+		"firebase": "^11.6.0",
+		"firebase-admin": "^13.3.0",
 		"html-to-text": "^9.0.5",
 		"html5-qrcode": "^2.3.8",
 		"imask": "^7.6.1",
-		"lodash-es": "^4.17.21",
-		"nodemailer": "^6.9.14",
+		"lodash.isstring": "^4.0.1",
+		"lodash.merge": "^4.6.2",
+		"nodemailer": "^6.10.1",
 		"pdf-lib": "^1.17.1",
-		"readable-web-to-node-stream": "^3.0.2"
+		"read-excel-file": "^5.8.7"
 	}
 }
