@@ -1,188 +1,105 @@
 <script lang="ts">
-	import { goto } from '$app/navigation';
-	import { page } from '$app/stores';
-	import { getAll } from '$lib/client/firestore';
-	import { nazevIR } from '$lib/Data';
-	import { relUrl } from '$lib/helpers/stores';
-	import { addToHistory, history, removeFromHistory, type HistoryEntry } from '$lib/History';
-	import type { Translations } from '$lib/translations';
-	import IMask from 'imask';
-	import { onMount } from 'svelte';
-	import { wordsToFilter } from '../new/companies';
+    import { goto } from '$app/navigation';
+    import { page } from '$app/state';
+    import { extractIRIDFromRawData, extractSPIDFromRawData, getAll, type IRID, publicProtocols, type SPID } from '$lib/client/firestore';
+    import type { Translations } from '$lib/translations';
+    import { onMount } from 'svelte';
+    import { p, SearchWidget } from '$lib/Widget.svelte.js';
+    import Search from '$lib/components/widgets/Search.svelte';
+    import { setTitle } from '$lib/helpers/title.svelte';
+    import { relUrl } from '$lib/helpers/runes.svelte';
+    import { nazevIR, nazevSP, popisIR } from '$lib/helpers/ir';
+    import { isUserRegulusOrAdmin } from '$lib/client/auth';
 
-	let nacita = true;
-	onMount(() => (nacita = false));
+    type Installation = {
+        irid: IRID,
+        label: string,
+        irName: string,
+    }
+    type PublicServiceProtocol = {
+        spid: SPID,
+        label: string,
+        id: string,
+    }
+    onMount(async () => {
+        const installationsSnapshot = await getAll();
+        const installations = installationsSnapshot.docs
+            .map(snapshot => snapshot.data())
+            .map(data => <Installation> {
+                irid: extractIRIDFromRawData(data.evidence),
+                irName: nazevIR(data.evidence.ir),
+                label: popisIR(data.evidence)
+            })
+            .filter(i => i.irid)
+            .toSorted((a, b) => a.irid.localeCompare(b.irid));
+        w1.items = () => installations;
 
-	let all: HistoryEntry[];
-	onMount(async () => {
-		const res = await getAll();
-		all = res.docs
-			.map((sn) => sn.data())
-			.map(
-				(data) =>
-					<HistoryEntry>{
-						ir: data.evidence.ir.cislo,
-						irType: nazevIR(t, data.evidence.ir.typ),
-						label: `${data.evidence.koncovyUzivatel.prijmeni} ${data.evidence.koncovyUzivatel.jmeno} ${data.evidence.mistoRealizace.obec}`
-					}
-			)
-			.toSorted((a, b) => a.ir.localeCompare(b.ir));
-	});
+        if ($isUserRegulusOrAdmin) {
+            const protocolsSnapshot = await publicProtocols();
+            const protocols = protocolsSnapshot.docs
+                .map(snapshot => snapshot.data())
+                .map(data => <PublicServiceProtocol> {
+                    spid: extractSPIDFromRawData(data.zasah),
+                    label: popisIR(data),
+                    id: nazevSP(data.zasah),
+                })
+                .filter(i => i.spid)
+                .toSorted((a, b) => -a.spid.localeCompare(b.spid));
+            w2.items = () => protocols;
+        }
+    });
 
-	$: filtered = !all
-		? []
-		: all.filter((entry) =>
-				wordsToFilter(search).every(
-					(filter) =>
-						wordsToFilter(entry.ir).join(' ').includes(filter) ||
-						wordsToFilter(entry.label).some((word) => word.startsWith(filter))
-				)
-			);
+    const t: Translations = page.data.translations;
 
-	const t: Translations = $page.data.translations;
+    let w1 = $state(new SearchWidget({
+        type: 'search',
+        required: false,
+        label: 'search',
+        items: [] as Installation[],
+        getSearchItem: i => ({
+            href: relUrl(`/detail/${i.irid}`),
+            pieces: [
+                { text: p`${i.irName}`, width: .4 },
+                { text: p`${i.label}`, width: .6 },
+            ] as const,
+        }),
+        onValueSet: (_, i) => {
+            if (i) goto(relUrl(`/detail/${i.irid}`));
+        }
+    }));
 
-	let input: HTMLInputElement | undefined;
-	$: mask =
-		input == undefined
-			? undefined
-			: IMask(input, {
-					mask: 'A1 0000',
-					definitions: {
-						A: /[A-Z]/,
-						'1': /[1-9OND]/
-					}
-				});
+    let w2 = $state(new SearchWidget({
+        type: 'search',
+        required: false,
+        label: 'search',
+        items: [] as PublicServiceProtocol[],
+        getSearchItem: pt => ({
+            href: relUrl(`/detail/${pt.spid}`),
+            pieces: [
+                { text: p`${pt.id}`, width: .4 },
+                { text: p`${pt.label}`, width: .6 },
+            ] as const,
+        }),
+        onValueSet: (_, i) => {
+            if (i) goto(relUrl(`/detail/${i.spid}`));
+        }
+    }));
 
-	onMount(() => {
-		return () => {
-			input = undefined;
-		};
-	});
-
-	let search = '';
-
-	$: mask?.on('accept', (_) => (search = mask.value));
+    setTitle(t.controllerSearch);
 </script>
 
-<h1>{t.controllerSearch}</h1>
-<form
-	on:submit|preventDefault={() => {
-		if (search == '' || filtered.length == 0) return;
-		const ir = filtered[0];
-		addToHistory(ir);
-		goto($relUrl(`/detail/${ir.ir.replace(' ', '')}`));
-	}}
->
-	<label class="form-floating d-block">
-		<input
-			type="search"
-			class="form-control border"
-			class:border-bottom-0={filtered.length > 0 && search != ''}
-			class:rb-0={filtered.length > 0 && search != ''}
-			placeholder={t.serialNumber}
-			bind:value={search}
-		/>
-		<label for="">{t.search}</label>
-	</label>
+<Search
+    bind:widget={w1}
+    data={undefined}
+    {t}
+/>
 
-	{#if search != '' && filtered.length > 0}
-		<div class="list-group">
-			{#each filtered as ir, i}
-				<a
-					class="list-group-item-action list-group-item d-flex flex-column flex-md-row flex-row align-items-md-center"
-					class:rt-0={i == 0}
-					href={$relUrl(`/detail/${ir.ir.replace(' ', '')}`)}
-					on:click|preventDefault={() => {
-						addToHistory(ir);
-						goto($relUrl(`/detail/${ir.ir.replace(' ', '')}`));
-					}}
-				>
-					<div class="col-md-5">{ir.irType} {ir.ir}</div>
-					<div class="col-md-7">{ir.label}</div>
-				</a>
-			{/each}
-		</div>
-	{/if}
-</form>
+{#if $isUserRegulusOrAdmin}
+    <h1>{t.protocolSearch}</h1>
 
-{#if $history.length != 0}
-	<h2 class="mt-2">Historie vyhledávání</h2>
-	<div class="list-group">
-		{#each $history.toReversed() as ir}
-			<div class="d-flex list-group-item-group flex-nowrap">
-				<a
-					class="list-group-item-action list-group-item d-flex flex-column flex-md-row flex-row align-items-md-center"
-					href={$relUrl(`/detail/${ir.ir.replace(' ', '')}`)}
-					on:click|preventDefault={() => {
-						addToHistory(ir);
-						goto($relUrl(`/detail/${ir.ir.replace(' ', '')}`));
-					}}
-				>
-					<div class="col-md-5">{ir.irType} {ir.ir}</div>
-					<div class="col-md-7">{ir.label}</div>
-				</a>
-				<button
-					class="btn btn-outline-danger rl-0"
-					on:click|stopImmediatePropagation|preventDefault={() => removeFromHistory(ir)}
-				>
-					<svg
-						xmlns="http://www.w3.org/2000/svg"
-						width="16"
-						height="16"
-						fill="currentColor"
-						viewBox="0 0 16 16"
-					>
-						<path
-							d="M2.146 2.854a.5.5 0 1 1 .708-.708L8 7.293l5.146-5.147a.5.5 0 0 1 .708.708L8.707 8l5.147 5.146a.5.5 0 0 1-.708.708L8 8.707l-5.146 5.147a.5.5 0 0 1-.708-.708L7.293 8z"
-						/>
-					</svg>
-				</button>
-			</div>
-		{/each}
-	</div>
+    <Search
+        data={undefined}
+        {t}
+        bind:widget={w2}
+    />
 {/if}
-
-<style>
-	.w-40 {
-		width: 40%;
-	}
-	.w-10 {
-		width: 10%;
-	}
-	.rb-0 {
-		border-bottom-left-radius: 0;
-		border-bottom-right-radius: 0;
-	}
-	.rt-0 {
-		border-top-left-radius: 0;
-		border-top-right-radius: 0;
-	}
-	.rl-0 {
-		border-top-left-radius: 0;
-		border-bottom-left-radius: 0;
-	}
-	.form-control:focus {
-		box-shadow: none;
-	}
-	.list-group-item-group:first-child .list-group-item {
-		border-top-left-radius: var(--bs-border-radius);
-	}
-	.list-group-item-group .list-group-item {
-		border-right: none;
-	}
-	.list-group-item-group:not(:last-child) .list-group-item {
-		border-bottom: none;
-	}
-	.list-group-item-group:last-child .list-group-item {
-		border-bottom-left-radius: var(--bs-border-radius);
-	}
-	.list-group-item-group:not(:last-child) .btn {
-		border-bottom: none;
-	}
-	.list-group-item-group:not(:last-child) .btn {
-		border-bottom-right-radius: 0;
-	}
-	.list-group-item-group:not(:first-child) .btn {
-		border-top-right-radius: 0;
-	}
-</style>
