@@ -1,47 +1,55 @@
 <script lang="ts">
 	import { browser } from '$app/environment';
-	import { page } from '$app/stores';
+	import { page } from '$app/state';
 	import type { Translations } from '$lib/translations';
 	import { onMount } from 'svelte';
 	import authentication from '$lib/client/authentication';
 	import FormDefaults from '$lib/components/FormDefaults.svelte';
-	
-	let nacita = true;
-	onMount(() => (nacita = false));
+	import { setTitle } from '$lib/helpers/title.svelte';
 
-	let email = browser ? $page.url.searchParams.get('email') ?? '' : '';
+	let odesila = $state(false);
+	let email = $state(browser ? (page.url.searchParams.get('email') ?? '') : '');
 
-	const t: Translations = $page.data.translations;
+	const t: Translations = page.data.translations;
 
 	let redirect: string = '/new';
-	onMount(() => (redirect = $page.url.searchParams.get('redirect') ?? '/new'));
+	onMount(() => (redirect = page.url.searchParams.get('redirect') ?? '/new'));
 
-	let error: string | null = null;
+	let error: string | null = $state(null);
 
 	const signUp = async () => {
+		odesila = true;
 		error = '';
 		const { enabled } = await authentication('checkEnabled', { email });
 		if (enabled) {
+			odesila = false;
 			error = t.emailInUse;
 			return;
 		}
 		if (enabled == null && email.endsWith('@regulus.cz')) {
-			await authentication('createUser', { email });
+			if (email.split('@')[0].includes('.'))
+				await authentication('createUser', { email });
+			else {
+				odesila = false;
+				error = 'Prosím, použijte email se jménem i příjmením';
+				return;
+			}
 		} else if (enabled == null) {
-			error = t.pleaseUseBuisnessEmail;
+			odesila = false;
+			error = t.pleaseUseBusinessEmail;
 			return;
 		}
 		const { link } = await authentication('getPasswordResetLink', {
 			email,
-			lang: $page.data.languageCode,
+			lang: page.data.languageCode,
 			redirect,
-			mode: 'register',
+			mode: 'register'
 		});
 		window.location.href = link;
 	};
-</script>
 
-<h1>{t.signUp}</h1>
+	setTitle(t.signUp)
+</script>
 
 <form>
 	<FormDefaults />
@@ -58,10 +66,14 @@
 		<p class="text-danger mt-3 mb-0">{@html error}</p>
 	{/if}
 	<div class="d-flex align-content-center mt-3">
-		<button type="submit" class="btn btn-primary me-2" on:click={signUp}>
-			{t.toSignUp}
-		</button>
-		<button type="button" class="btn btn-outline-secondary" on:click={() => history.back()}>
+		{#if odesila}
+			<div class="spinner-border text-danger m-2"></div>
+		{:else}
+			<button type="submit" class="btn btn-primary me-2" onclick={signUp}>
+				{t.toSignUp}
+			</button>
+		{/if}
+		<button type="button" class="btn btn-secondary" onclick={() => history.back()}>
 			{t.back}
 		</button>
 	</div>
