@@ -1,28 +1,70 @@
-import type { Options } from "nodemailer/lib/mailer";
-import { getToken } from "./auth";
-import { htmlToText } from "html-to-text";
-import { isString } from "lodash-es";
+import type { Address, Attachment } from 'nodemailer/lib/mailer';
+import { currentUser, getToken } from './auth';
+import { htmlToText } from 'html-to-text';
+import { type Component, mount } from 'svelte';
+import { dev } from '$app/environment';
+import { get } from 'svelte/store';
+import type { User } from 'firebase/auth';
 
-/**
- * Pošle email a pokud tam je html, převede to na text
- */
+export type AddressLike = Address | string | (Address | string)[];
+export type EmailOptions<Props extends Record<string, unknown>> = {
+    from: Address | string;
+    replyTo?: AddressLike;
+    to?: AddressLike;
+    cc?: AddressLike;
+    bcc?: AddressLike;
+    subject: string;
+    attachments?: Attachment[];
+    pdf?: {
+        link: string;
+        title: string;
+    };
+    component: Component<Props, Record<string, unknown>, '' | keyof Props>;
+    props: Props;
+}
+export type EmailMessage = Omit<EmailOptions<never>, 'component' | 'props'> & {
+    text: string, html: string,
+}
 
-export const sendEmail = async (message: Options) => {
+export const sendEmail = async <Props extends Record<string, unknown>>(options: EmailOptions<Props>) => {
+    const div = document.createElement('div');
+    mount(options.component, {
+        target: div,
+        props: options.props,
+    });
+    const html = div.innerHTML;
+    const message: EmailMessage = {
+        ...options, html, text: htmlToText(html)
+    };
 
-	if (isString(message.html)) {
-		message.text = message.text ?? htmlToText(message.html);
-	}
-
-	const token = await getToken();
-
-	return await fetch(`/api/sendEmail?token=${token}`, {
-		method: 'POST',
-		body: JSON.stringify({ message }),
-		headers: {
-			'content-type': 'application/json'
-		}
-	});
+    const token = await getToken();
+    return await fetch(`/api/sendEmail?token=${token}`, {
+        method: 'POST',
+        body: JSON.stringify({ message }),
+        headers: {
+            'content-type': 'application/json'
+        }
+    });
 };
 
-export const SENDER = '"Regulus Evidence IR" aplikace.regulus@centrum.cz';
+export const cervenka: AddressLike = ['david.cervenka@regulus.cz', 'jakub.cervenka@regulus.cz']
 
+export const SENDER: Address = {
+    name: 'Regulus SEIR',
+    address: 'aplikace.regulus@gmail.com',
+};
+
+const userAddress = (user: User): AddressLike => ({
+    address: user.email!,
+    name: user.displayName ?? '',
+});
+
+export const defaultAddresses = (recipient: AddressLike = cervenka, sendCopy: boolean = false) => {
+    const user = userAddress(get(currentUser)!);
+    return ({
+        from: SENDER,
+        replyTo: user,
+        to: dev ? 'radek.blaha.15@gmail.com' : recipient,
+        cc: dev || !sendCopy ? undefined : user,
+    });
+}
