@@ -1,10 +1,10 @@
 <script>
 	import { browser } from '$app/environment';
-	import { page } from '$app/stores';
+	import { page } from '$app/state';
 </script>
 
 {#if browser}
-	{#each $page.url.searchParams.toString().split('&') as param}
+	{#each page.url.searchParams.toString().split('&') as param}
 		<input
 			type="hidden"
 			name={param.split('=')[0]}
