@@ -1,11 +1,11 @@
 import { get } from "svelte/store"
 import { storable } from './helpers/stores'
 
-export const languageCodes = ["cs", "en", "de"/* , "sk" */] as const
+export const languageCodes = ["cs", "en", "de", "sk" ] as const
 
-export const isLanguageCode = (code: any): code is LanguageCode => languageCodes.includes(code)
+export const isLanguageCode = (code: unknown): code is LanguageCode => (languageCodes as readonly unknown[]).includes(code)
 
-export const asLanguageCodeOrNull = (code: any): LanguageCode | null =>
+export const asLanguageCodeOrNull = (code: unknown): LanguageCode | null =>
     isLanguageCode(code) ? code : null
 
 export type LanguageCode = typeof languageCodes[number]
@@ -14,7 +14,7 @@ export const defaultLanguage: LanguageCode = "en"
 
 const localLanguage: () => LanguageCode | undefined = () => navigator.languages.find(it => isLanguageCode(it))
 
-const userPreferedLanguage = storable<LanguageCode | undefined>(undefined, "userPreferedLanguage")
+const userPreferredLanguage = storable<LanguageCode>("user_preferred_language")
 
-export const setUserPreferedLanguage = (code: LanguageCode) => userPreferedLanguage.set(code)
-export const preferedLanguage: () => LanguageCode = () => get(userPreferedLanguage) ?? localLanguage() ?? defaultLanguage
+export const setUserPreferredLanguage = (code: LanguageCode) => userPreferredLanguage.set(code)
+export const preferredLanguage: () => LanguageCode = () => get(userPreferredLanguage) ?? localLanguage() ?? defaultLanguage
