@@ -1,51 +1,86 @@
-import { addParsing, type AddParsing } from "./helpers/templates";
-import type { LanguageCode } from "./languages";
+import { addParsing, type AddParsing, type STemplate, type Template, type TemplateArgs, type Templates } from './helpers/templates';
+import type { LanguageCode } from './languages';
 import cs from './translations/cs';
 import de from './translations/de';
 import en from './translations/en';
-import sk from './translations/sk';
-import { createTemplateG } from "./Vec";
+import sk from '$lib/translations/sk';
+import derived, { type Derived } from '$lib/derived';
+import './extensions';
+import merge from 'lodash.merge';
 
-const translationsMap: PlainTranslationsMap = { cs, en, de/*, sk */ };
+const translationsMap: PlainTranslationsMap = { cs, en, de, sk };
+
+export const makePlain = <T extends string | undefined | null>(text: T) => text ? `PLAIN_${text}` as const : undefined;
+export const removePlain = (ref: string) => ref.slice(6);
+export const isPlain = (ref: TranslationReference | TemplateKey) => ref.startsWith('PLAIN_');
+
+type Translation = string | Template<(string | number)[]>;
 
 const withGet = (translations: PlainTranslations): Translations => {
-    const get = (ref: TranslationReference) => ref == '' ? '' : ref.startsWith('PLAIN_') ? ref.slice(6) : ref.split('.').reduce<Record<string, any> | string>((acc, key) => (acc as Record<string, any>)[key], translations) as string ?? ref
-    return {
-        ...addParsing(translations),
-        get,
-        getN: (ref) => ref == null ? ref : get(ref),
-        getT: createTemplateG(({ strings, args }) => ({ strings, args: args.map(get) }), a => a)
-    }
-}
+    const withParsing = addParsing(translations) as AddParsing<PlainTranslations>;
+    const withDerived = merge<AddParsing<PlainTranslations>, Derived>(withParsing, derived(withParsing));
+    const get = (ref: TranslationReference | TemplateKey) => ref == '' ? ''
+        : isPlain(ref) ? removePlain(ref) : ((ref
+            .split('.')
+            .reduce<Record<string, unknown> | Translation>((acc, key) =>
+                (acc as Record<string, Translation | Record<string, unknown>>)[key], withDerived
+            ) as Translation) ?? ref);
+    return <Translations> {
+        ...withDerived,
+        get: ref => ref == null ? null : get(ref),
+        refFromTemplate: <T extends (number | string)[]>(ref: TemplateKey, args: TemplateArgs<T>) =>
+            `PLAIN_${(get(ref) as Template<T>).parseTemplate(args)}`,
+    };
+};
 
-export type PlainTranslations = typeof cs
+export type PlainTranslations = typeof cs;
 
+type TemplateKey = RecursiveKeyOf<Templates<PlainTranslations>>;
 export type Translations = AddParsing<PlainTranslations> & {
-    get: (ref: TranslationReference) => string,
-    getN: (ref: TranslationReference | null) => string | null,
-    getT: (strings: TemplateStringsArray, ...args: TranslationReference[]) => string,
-}
-export type TranslationReference = RecursiveKeyof<PlainTranslations> | `PLAIN_${string}` | ''
+    get: <T extends TranslationReference | TemplateKey | null>(ref: T) => T extends null ? null : T extends TemplateKey ? Template<(string | number)[]> : string;
+    refFromTemplate: <K extends TemplateKey>(
+        ref: K,
+        args: TemplateArgs<AtRecursiveKey<PlainTranslations, K> extends STemplate<infer T> ? T : never>
+    ) => TranslationReference
+} & Derived;
+
+export const allKeys = <(TranslationReference | TemplateKey)[]> Object.recursiveKeys(cs);
+
+export type TranslationReference = Exclude<
+    | RecursiveKeyOf<PlainTranslations & Derived>
+    | `PLAIN_${string}`
+    | '',
+    TemplateKey>;
 
 export type Translate<T> = {
-    [Code in LanguageCode]: T
-}
+    [Code in LanguageCode]: T;
+};
 
-type PlainTranslationsMap = Translate<PlainTranslations>
+type PlainTranslationsMap = Translate<PlainTranslations>;
 
-export const getTranslations = (code: LanguageCode): Translations =>
-    withGet(translationsMap[code])
+export const getTranslations = (code: LanguageCode): Translations => withGet(translationsMap[code]);
 
-type LanguageNames = Translate<string>
+type LanguageNames = Translate<string>;
 
 export const languageNames: LanguageNames = {
-    cs: "čeština",
-    en: "English",
-//     sk: "slovenčina",
-    de: "Deutsch",
-}
+    cs: 'čeština',
+    en: 'English',
+    sk: 'slovenčina',
+    de: 'Deutsch'
+};
 
-type RecursiveKeyof<T extends object> = T extends string ? T : {
-    //@ts-expect-error
-    [K in keyof T]: T[K] extends object ? `${K extends string ? K : ''}.${RecursiveKeyof<T[K]> extends string ? RecursiveKeyof<T[K]> : ''}` : K
-}[keyof T]
+type RecursiveKeyOf<T extends Record<string, unknown>> = T extends string
+    ? T
+    : {
+        [K in keyof T]: T[K] extends Record<string, unknown>
+            ? RecursiveKeyOf<T[K]> extends infer Key
+                ? `${K extends string ? K : ''}.${Key extends string ? Key : ''}`
+                : never
+            : K;
+    }[keyof T];
+
+type AtRecursiveKey<T extends Record<string, unknown>, K extends RecursiveKeyOf<T>> =
+    K extends `${infer K1}.${infer K2}`
+        ? T[K1] extends Record<string, unknown> ?
+            K2 extends RecursiveKeyOf<T[K1]> ? AtRecursiveKey<T[K1], K2> : never : never
+        : T[K]
