@@ -1,30 +1,43 @@
-import { template as t } from "$lib/helpers/templates"
+/* eslint-disable no-irregular-whitespace */
+import { template as t } from '$lib/helpers/templates';
 
 export default {
     controllerRegistration: `Registrace nového regulátoru`,
-    longAppName: `Evidence regulátorů IR`,
-    appName: `Evidence IR`,
+    longAppName: `SEIR – Evidence regulátorů IR`,
+    appName: `SEIR`,
     requiredField: `Toto pole je povinné`,
     controllerType: `Typ regulátoru`,
     serialNumber: `Sériové číslo regulátoru`,
     serialNumberIndoor: `Sériové číslo vnitřní jednotky`,
+    recognised: t`Rozpoznáno: ${0}`,
     newSerialNumber: `Nové sériové číslo`,
     wrongNumberFormat: `Nesprávný formát čísla`,
     heatPumpType: `Typ tepelného čerpadla`,
+    heatPumpsType: `Typ tepelných čerpadel`,
     airToWater: `vzduch/voda`,
     groundToWater: `země/voda`,
-    heatPumpModel: `Model tepelného čerpadla`,
-    heatPumpManufactureNumber: `Výrobní číslo tepelného čerpadla`,
+    heatPumpModelNr: t`Model ${0}tepelného čerpadla`,
+    noPump: `Žádný`,
+    first: `prvního`,
+    second: `druhého`,
+    third: `třetího`,
+    fourth: `čtvrtého`,
+    heatPumpManufactureNumberNr: t`Výrobní číslo ${0} tepelného čerpadla`,
     endUser: `Koncový uživatel`,
+    individual: `Jednotlivec (FO)`,
+    company: `Společnost (PO)`,
     name: `Jméno`,
     surname: `Příjmení`,
     birthday: `Datum narození`,
+    companyName: `Název`,
+    wrongCompanyType: `Pozor, zadaná forma společnosti není správně formátovaná!`,
+    establishment: `Pobočka/závod`,
     wrongDateFormat: `Nesprávný formát datumu`,
     phone: `Telefon`,
     wrongPhoneFormat: `Nesprávný formát telefoního čísla`,
     email: `Email`,
     wrongEmailFormat: `Nesprávný formát emailu`,
-    roalizationLocation: `Místo realizace`,
+    realizationLocation: `Místo realizace`,
     town: `Obec`,
     street: `Číslo popisné nebo ulice a číslo orientační`,
     zip: `Poštovní směrovací číslo`,
@@ -34,6 +47,7 @@ export default {
     wrongCRNFormat: `Nesprávný formát IČO`,
     representativeName: `Jméno zástupce`,
     commissioning: `Uvedení do provozu`,
+    dateOfCommission: `Datum uvedení`,
     commissionedByAssemblyCompany: `Do provozu uváděla montážní firma`,
     remoteAccess: `Vzdálený přístup`,
     doYouWantRemoteAccess: `Založit vzdálený přístup k regulátoru`,
@@ -44,6 +58,7 @@ export default {
     doNotInvoice: `Nefakturovat`,
     responsiblePerson: `Zodpovědná osoba`,
     chosenCompany: `Vybraná firma`,
+    chosenFile: `Vybráno:`,
     cancel: `Zrušit`,
     password: `Heslo`,
     confirmPassword: `Potvrdit heslo`,
@@ -55,24 +70,27 @@ export default {
     wrongPasswordHtml: t`Špatné heslo! <a href="${'link'}">Zapomenuté heslo?</a>`,
     tooManyRequests: `Moc žádostí! Počkejte prosím chvíli`,
     somethingWentWrong: `Něco se nepovedlo :\\`,
+    somethingWentWrongContactUsHtml: `Omlouváme se, něco se nepovedlo, kontaktujte nás prosím na <a href="mailto:Regulus SEIR<aplikace.regulus@gmail.com>?subject=Chyba při ukládání formuláře" target="_blank">aplikace.regulus@gmail.com</a>`,
     passwordTooWeak: `Heslo je příliš slabé!`,
-    pleaseUseBuisnessEmail: `Prosím zadejte Váš firemní email`,
+    pleaseUseBusinessEmail: `Prosím zadejte Váš firemní email`,
     emailInUse: `Tento účet již existuje`,
     passwordsDoNotMatch: `Hesla se neshodují!`,
     save: `Uložit`,
+    saveAndSend: `Uložit a odeslat emaily`,
     saving: `Ukládání...`,
-    youHaveAMistake: `Ve formuláři máte chybu. Prosím, zkontrolujte vyplněné informace.`,
+    youHaveAMistake: t`Ve formuláři máte chybu. Prosím, zkontrolujte vyplněné informace. Chybná pole: ${'fields'}`,
     logInNeeded: `Pro zobrazení a vyplnění formuláře je nutné se přihlásit!`,
     evidenceDetailsHtml: t`Podrobnosti o instalaci s ${'irType'} ${'irNumber'}`,
     evidenceDetails: `Podrobnosti o instalaci`,
     loadingData: `Načítání dat...`,
     sorrySomethingWentWrong: `Omlouváme se, něco se nepovedlo.`,
-    linkInvalid: `Buď je záznam o evidenci neexistuje nebo k němu nemáte přístup.`,
+    offline: `Jste offline!`,
+    linkInvalid: `Buď záznam o evidenci neexistuje nebo k němu nemáte přístup.`,
     linkToThis: `Odkaz na tuto stránku`,
     copy: `Kopírovat`,
-    regulusRouteForm: `Formulář o zpřístupnění regulátoru službě RegulusRoute`,
+    regulusRouteForm: `Souhlas s RegulusRoute`,
     regulusRouteTitle: `Souhlas se zpřístupněním regulátoru IR službě RegulusRoute`,
-    regulusRouteFileName: `Formulář RegulusRoute.pdf`,
+    regulusRouteFileName: `Souhlas s RegulusRoute.pdf`,
     openPdf: `Otevřít pdf`,
     changeLang: `Změnit jazyk`,
     deleteThisEvidence: `Odstranit tento záznam evidence`,
@@ -87,30 +105,40 @@ export default {
     scanBarcode: `Naskenujte čárový kód z TČ`,
     scanCode: `Naskenujte kód`,
     residence: `Bydliště`,
+    headquarters: `Sídlo`,
     whatToAddInfoTo: `K čemu byste chtěli přidat podrobnější informace?`,
     heatPump: `Tepelné čerpadlo`,
+    heatPumps: `Tepelná čerpadla`,
+    solarCollector: `Solární kolektor`,
+    solarSystem: `Solární systém`,
+    solarCollectorType: `Typ solárního kolektoru`,
+    solarCollectorCount: `Počet solárních kolektorů`,
     pleaseFillInIrType: `Prosím, vyplňte typ IR.`,
     fillInPassword: `Prosím, zadejte heslo.`,
     search: `Vyhledat`,
     controllerSearch: `Vyhledávání instalací`,
+    protocolSearch: `Vyhledávání servisních protokolů`,
     new: `Nová registrace`,
     changeController: `Změnit sériové číslo regulátoru`,
     confirm: `Potvrdit`,
     changeWentWrong: `Něco ne nepovedlo. Obnovte stránku a zkuste to znovu.`,
     note: `Poznámka`,
     routeGuide: `Návod na přihlášení k regulátoru`,
-    warranty: `Záruční list k TČ`,
-    commissionProtocol: `Protokol o instalaci TČ`,
-    commissionProtocolTitle: `Protokol o uvedení tepelného čerpadla do trvalého provozu`,
-    commissionProtocolFileName: `Protokol uvedení TČ`,
-    installationApproval: `Schválení instalace TČ`,
-    yearlyCheck: `Roční kontrola TČ`,
+    warrantyNr: t`Záruční list TČ${0}`,
+    heatPumpCommissionProtocol: `Protokol o instalaci TČ`,
+    heatPumpCommissionProtocolTitle: `Protokol o uvedení tepelného čerpadla do trvalého provozu`,
+    heatPumpCommissionProtocolFileName: `Protokol uvedení TČ`,
+    solarCollectorCommissionProtocol: `Protokol o instalaci solárního systému`,
+    solarCollectorCommissionProtocolFileName: `Protokol uvedení SOL`,
+    yearlyHPCheck: `Roční kontrola TČ`,
     yearlyCheckFileName: `Roční prohlídka.pdf`,
     yearlyCheckTitle: `Popis úkonů při provádění preventivní roční prohlídky vzduchového tepelného čerpadla`,
     doYearlyCheck: `Udělat roční kontrolu TČ`,
+    checkDate: `Datum kontroly`,
     filledYearlyCheck: `Roční kontroly TČ`,
-    chooseFromList: `Vyberte ze seznamu`,
-    chooseCompanyFromList: `Vyberte firmu ze seznamu:`,
+    searchCompanyInList: `Vyhledat firmu v seznamu`,
+    searchRepresentative: `Vyhledat zástupce`,
+    representative: `Zástupce`,
     newPassword: `Nové heslo`,
     sendConfirmEmail: `Odeslat potvrzovací email`,
     passwordReset: `Obnovení hesla`,
@@ -126,6 +154,7 @@ export default {
     editRegistration: `Upravit údaje o instalaci`,
     requiresLogIn: `Tato stránka vyžaduje přihlášení.`,
     emptyForm: `Vymazat data`,
+    importData: `Importovat data`,
     siteNotFound: `Stránka nenalezena.`,
     changePassword: `Změnit heslo`,
     year: `Rok`,
@@ -140,7 +169,8 @@ export default {
     hpWarranty: `Záruční list tepelného čerpadla`,
     warrantyFileName: `Záruční list.pdf`,
     performingPerson: `Provádějící osoba`,
-    realisedAtResidence: `Shodné s bydlištěm`,
+    samePlaceAsResidence: `Shodné s bydlištěm`,
+    samePlaceAsHeadquarters: `Shodné se sídlem`,
     agreeWIthRRPrice: `Souhlasím s jednorázovou cenou 1000 Kč včetně DPH za tuto službu.`,
     suits: `vyhovuje`,
     suitsNot: `nevyhovuje`,
@@ -192,7 +222,291 @@ export default {
     wasControllerSetToParameters: `Regulátor tepelného čerpadla byl nastaven na předepsané parametry`,
     wasOwnerFamiliarizedWithFunction: `Vlastník nebo provozovatel byl seznámen se základní funkcí tep. čerpadla a jeho obsluhou`,
     isExtendedWarrantyDesired: `Má vlastník TČ zájem o prodlouženou záruku?`,
-    extendedWarranty7Years: `ano – 7 let na kompresor`,
-    extendedWarranty10Years: `ano – 10 let na kompresor (příplatek)`,
     isInstallationInWarrantyConditions: `Instalace a uvedení do provozu jsou v souladu s podmínkami prodloužené záruky`,
-}
+    successfullyDeleted: `Záznam úspěšně odstraněn`,
+    mandatoryFields: `povinná pole`,
+    cascadeSee: `kaskáda – viz níže`,
+    cascade: `Kaskáda:`,
+    pumpDetails: t`TČ${'n'}: ${'model'} – ${'cislo'}`,
+    theme: {
+        dark: 'Tmavý',
+        light: 'Světlý',
+        auto: 'Podle prohlížeče',
+    },
+    sp: {
+        yearlyCheck: `Roční prohlídka`,
+        warrantyRepair: `Záruční oprava`,
+        postWarrantyRepair: `Pozáruční oprava`,
+        installationApproval: `Schválení instalace`,
+        otherType: `Jiné`,
+        regulusRoute: `RegulusRoute`,
+        extendedWarranty: `Prodloužená záruka`,
+        commissioningTC: `Uvedení TČ do provozu`,
+        commissioningSOL: `Uvedení SOL do provozu`,
+        yearlySOLCheck: `Roční kontrola SOL`,
+        technicalAssistance: `Technická pomoc`,
+        assemblyWork: `Montážní práce`,
+    },
+    check: {
+        kontrolniElektroinstalace: 'Kontrolní elektroinstalace',
+        kontrolniUkonyTepelnehoCerpadla: 'Kontrolní úkony tepelného čerpadla',
+        kontrolniUkonyRegulace: 'Kontrolní úkony regulace',
+        kontrolniUkonyOtopneSoustavy: 'Kontrolní úkony otopné soustavy',
+        kontrolaZasobnikuTv: 'Kontrola zásobníku TV',
+        kontrolaFunkceVsechElektrickychSpotrebicuZapojenychDoRegulace: 'Kontrola funkce všech elektrických spotřebičů zapojených do regulace',
+        kontrolaDotazeniSvorkovychSpoju: 'Kontrola dotažení svorkových spojů',
+        vizualniKontrolaVsechPristupnychVodicuVInstalaciNataveniMechPoskozeni: 'Vizuální kontrola všech přístupných vodičů v instalaci (natavení, mech.poškození)',
+        kontrolaSepnutiDohrevuSepnutiStykacePripadneRele: 'Kontrola sepnutí dohřevu (sepnutí stykače, případně relé)',
+        kontrolaChoduKompresoru: 'Kontrola chodu kompresoru',
+        optickaKontrolaTesnostiTrubkovychSpojuJednotkyAChladivovehoOkruhu: 'Optická kontrola těsnosti trubkových spojů jednotky a chladivového okruhu',
+        kontrolaOdvoduKondenzatu: 'Kontrola odvodu kondenzátu',
+        kontrolaUchyceniVentilatoru: 'Kontrola uchycení ventilátoru',
+        vycisteniVzduchovychCestJednotky: 'Vyčištění vzduchových cest jednotky',
+        kontrolaLamelVyparnikuPripadneOdstraneniNecistotVzduchem: 'Kontrola lamel výparníku, případné odstranění nečistot vzduchem',
+        proveritZdaNicNebraniOptimalniCirkulaciVzduchu: 'Prověřit, zda nic nebrání optimální cirkulaci vzduchu',
+        kontrolaTeplotnihoRozdiluTepelnehoCerpadlaDleNavodu: 'Kontrola teplotního rozdílu tepelného čerpadla dle návodu',
+        kontrolaElektrickeCastiJednotkyTepelnehoCerpadla: 'Kontrola elektrické části jednotky tepelného čerpadla',
+        kontrolaChybovychAInformacnichHlaseniRegulatoruAJejichPricin: 'Kontrola chybových a informačních hlášení regulátoru a jejich příčin',
+        kontrolaNastaveniParametruRegulatoru: 'Kontrola nastavení parametrů regulátoru',
+        preventivniProskoleniObsluhyZHlediskaUzivatelskehoNastaveni: 'Preventivní proškolení obsluhy z hlediska uživatelského nastavení',
+        stavPocitadlaCelkovychProvoznichHodinKompresoru: 'Stav počítadla celkových provozních hodin kompresoru',
+        stavPocitadlaProvoznichHodinDoTvUmoznujeLiToRegulace: 'Stav počítadla provozních hodin do TV (umožňuje-li to regulace)',
+        stavCelkovehoPoctuStartuTepCerpadlaUmoznujeLiToRegulace: 'Stav celkového počtu startů tep. čerpadla (umožňuje-li to regulace)',
+        stavPoctuStartuTepelCerpadlaDoTvUmoznujeLiToRegulace: 'Stav počtu startů tepel. čerpadla do TV (umožňuje-li to regulace)',
+        stavPocitadlaCelkovychProvoznichHodinDoplnkovehoZdroje: 'Stav počítadla celkových provozních hodin doplňkového zdroje',
+        stavPocitadlaCelkovychProvoznichHodinDoplnkovehoZdrojeTv: 'Stav počítadla celkových provozních hodin doplňkového zdroje TV',
+        prumernaCelkovaDobaChoduKompresoruMinOdPosledniKontroly: 'Průměrná celková doba chodu kompresoru [min] - od poslední kontroly',
+        prumernaDobaChoduKompresoruDoTvMinOdPosledniKontroly: 'Průměrná doba chodu kompresoru do TV [min] - od poslední kontroly',
+        kontrolaFunkceObehovychCerpadel: 'Kontrola funkce oběhových čerpadel',
+        vycisteniFiltruObehovychCerpadel: 'Vyčištění filtrů oběhových čerpadel',
+        odvzdusneniZdrojeTc: 'Odvzdušnění zdroje (TČ)',
+        kontrolaFunkceVsechMotorickychVentiluSmesovaciZonovych: 'Kontrola funkce všech motorických ventilů (směšovací, zónových)',
+        kontrolaTesnostiOtopneSoustavy: 'Kontrola těsnosti otopné soustavy',
+        kontrolaTlakuVExpanzniNadobeOtopneSoustavy: 'Kontrola tlaku v expanzní nádobě otopné soustavy',
+        nastavenyTlakPriUvadeniDoProvozu: 'Nastavený tlak při uvádění do provozu',
+        pripadneProvedteKontroluTlakuVOtopneSoustave: 'Případně proveďte kontrolu tlaku v otopné soustavě',
+        kontrolaMgAnodyVZasobnikuPripVymena: 'Kontrola Mg anody v zásobníku, příp. výměna',
+        kontrolaPojistovacihoVentilu: 'Kontrola pojišťovacího ventilu',
+        pripadneProvedteKontroluTlakuVEnTepleVody: 'Případně proveďte kontrolu tlaku v EN teplé vody',
+    },
+    demand: {
+        demandForm: 'Dotazník poptávky',
+        contacts: {
+            contacts: 'Kontakty a místo realizace',
+            surname: 'Příjmení',
+            name: 'Jméno',
+            street: 'Ulice, č. p.',
+            city: 'Město',
+            zip: 'PSČ',
+            phone: 'Telefon',
+            email: 'Email',
+            chooseAssemblyCompanyHere: 'Vyberte montážní firmu:',
+            chooseCompany: 'Vybrat firmu',
+            chooseAssemblyCompany: 'Vyberte montážní firmu',
+            searchCompany: 'Vyhledat firmu',
+            crn: 'IČO monážní firmy',
+            demandOrigin: 'Původ poptávky',
+            originQuestionEmail: 'Dotaz poslaný emailem od koncového zákazníka',
+            originQuestionExhibition: 'Dotaz z výstavy',
+            originQuestionInPerson: 'Dotaz při osobní návštěvě u koncového zákazníka',
+            originDistributionCompany: 'Poptávka od distribuční firmy',
+            originAssembleres: 'Poptávka od montážníků',
+            originDesigner: 'Poptávka od projektanta',
+            demandSubject: 'Předmět poptávky',
+            heatPump: 'Tepelné čerpadlo',
+            fve: 'FVE',
+        },
+        objectDetail: {
+            objectDetail: 'Detail objektu',
+            heatLoss: 'Tepelná ztráta',
+            heatNeedsForHeating: 'Potřeba tepla – vytápění',
+            heatNeedsForHotWater: 'Potřeba tepla – teplá voda',
+            area: 'Vytápěná plocha',
+            volume: 'Vytápěný objem',
+            costs: 'Náklady na vytápění',
+            type: 'Druh paliva',
+            usage: 'Spotřeba paliva',
+            type2: 'Druh 2. paliva',
+            usage2: 'Spotřeba 2. paliva',
+        },
+        fve: {
+            fve: 'FVE',
+            currentSituation: 'Současný stav',
+            currentHeating: 'Současný zdroj vytápění',
+            currentHotWater: 'Současný zdroj přípravy teplé vody',
+            currentTanks: 'Instalované nádrže/zásobníky',
+            currentConsumption: 'Aktuální roční spotřeba el. energie',
+            breakerSize: 'Velikost hlavního jističe',
+            tariff: 'Sazba',
+            breakerBoxLocation: 'Umístění domovního rozvaděče',
+            requirements: 'Požadavky na FVE',
+            requiredPower: 'Požadovaný výkon',
+            locationBuidingType: 'Typ budovy pro instalaci kolektorů',
+            lightningRod: 'Na střeše pro instalaci je umístěn hromosvod',
+            familyHouseEtc: 'Rodinný dům, stodola, pergola, …',
+            avaiableAreas: 'Dostupné plochy pro instalaci',
+            roofMaterial: 'Materiál střesní krytiny',
+            roofAge: 'Stáří střešní krytiny',
+            tile: 'Taška',
+            tileType: 'Typ tašek',
+            metalSheetFolded: 'Plech falcovaný',
+            metalSheetTrapezoidal: 'Plech trapézový',
+            foil: 'Fólie',
+            asphaltShingle: 'Asfaltový šindel',
+            useOptimizers: 'Použít optimizéry',
+            size: t`Rozměr ${0}. plochy`,
+            orientation: t`Orientace ${0}. plochy`,
+            slope: t`Sklon ${0}. plochy`,
+            battery: 'Akumulace do baterií',
+            batteryCapacity: 'Kapacita baterií',
+            water: 'Akumulace do vody',
+            network: 'Přetoky do sítě',
+            networkPower: 'Přetoky do sítě – rezervovaný výkon',
+            charging: 'Dobíjecí stanice – wallbox',
+        },
+        system: {
+            system: 'Systém',
+            hPType: 'Typ tepelného čerpadla',
+            hPModel: 'Model tepelného čerpadla',
+            storeType: 'Typ nádrže',
+            storeNone: 'Žádná',
+            storeVolume: 'Objem nádrže',
+            tankType: 'Typ zásobníku',
+            tankNone: 'Žádná',
+            tankVolume: 'Objem zásobníku',
+            indoorUnitType: 'Typ vnitřní jednotky',
+            indoorUnitNone: 'Žádná',
+            heatingSystem: 'Otopný systém',
+            heatingSystem1circuit: '1 otopný okruh',
+            heatingSystem2circuits: '2 otopné okruhy',
+            heatingSystem3circuits: '3 otopné okruhy',
+            heatingSystemInvertor: 'Invertor na přímo',
+            heatingSystemOther: 'Jiný',
+            hotWaterCirculation: 'Cirkulace teplé vody',
+            poolHeating: 'Ohřev bazénu',
+            airToWater: 'vzduch/voda',
+            groundToWater: 'země/voda',
+        },
+        pool: {
+            pool: 'Bazén',
+            usagePeriod: 'Doba využívání',
+            periodYearlong: 'celeroční',
+            periodSeasonal: 'sezónní',
+            location: 'Umístění',
+            locationOutdoor: 'venkovní',
+            locationIndoor: 'vnitřní',
+            waterType: 'Druh vody',
+            freshType: 'sladká',
+            saltType: 'slaná',
+            shape: 'Tvar',
+            shapeRectangle: 'obdélníkový',
+            shapeOval: 'oválný',
+            shapeCircle: 'kruhový',
+            length: 'Délka',
+            width: 'Šířka',
+            radius: 'Průměr',
+            depth: 'Hloubka',
+            coverage: 'Zakrytí',
+            coverageNone: 'žádné',
+            coverageSolid: 'pevná střecha',
+            coveragePolycarbonate: 'polykarbonát',
+            coverageOther: 'jiné',
+            temperature: 'Požadovaná teplota',
+        },
+        additionalSources: {
+            sources: 'Doplňkové zdroje',
+            hotWater: 'Teplá voda',
+            heating: 'Topení',
+            heatingElement: 'Topné těleso v nádrži',
+            electricBoiler: 'Elektrokotel',
+            gasBoiler: 'Plynový kotel',
+            fireplace: 'Krb nebo kotel na tuhá paliva',
+            newNeuter: 'nové',
+            newMasculine: 'nový',
+            existing: 'stávající',
+            toSocket: 'do zásuvky',
+            fromRegulation: 'ovládané z regulace',
+            otherSource: 'Jiný',
+        },
+        accessories: {
+            accessories: 'Příslušenství',
+            hose: 'Hadice',
+            wallSupportBracket: 'Držák pro TČ',
+            roomUnitsAndSensors: 'Pokojové jednotky a čidla',
+            heatingCable: 'Topný kabel',
+            onWall: 'Na stěnu',
+            onIsolatedWall: 'Na izolovanou stěnu',
+        },
+        logIn: {
+            logIn: 'Přihlášení',
+            internetNeeded: 'Pro přihlášení je potřeba připojení k internetu!',
+            iAmEmploee: 'Jsem zaměstanec Regulusu',
+            iAmNotEmploee: 'Nejsem zaměstanec Regulusu',
+            yourRepresentative: 'Váš obchodní zástupce',
+            yourName: 'Vaše jméno',
+            yourSurname: 'Vaše příjmení',
+            yourCrn: 'Vaše IČO (nepovinné)',
+            yourEmail: 'Váš email',
+            chooseYourself: 'Vyber se:',
+            selectedFullName: t`Jméno a příjmení: ${'jmeno'} ${'prijmeni'}`,
+            selectedEmail: t`Email: ${'email'}`,
+            selectedCode: t`Číslo KO: ${'kod'}`,
+            selectedCrn: t`IČO: ${'ico'}`,
+            nameAndSurname: 'Jméno a příjmení:',
+            email: 'Email:',
+            code: 'Číslo KO:',
+            crn: 'IČO:',
+            addMoreInfo: 'Doplntě, prosím, ještě nějaké informace o Vás:',
+            youAreNotLoggedIn: 'Nejste přihlášeni!',
+            logOut: 'Odhlásit se',
+            representativeNeeded: 'Je potřeba vybrat obchodního zástupce',
+            emailNeeded: 'Je potřeba vyplnit email',
+            surnameNeeded: 'Je potřeba vyplnit příjmení',
+            nameNeeded: 'Je potřeba vypllnit jméno',
+        },
+        photos: {
+            photosManager: 'Správa fotek',
+            takePhoto: 'Vyfotit',
+            choosePhoto: 'Vybrat z galerie',
+            photo: 'Fotka',
+            add: 'Přidat fotku',
+            maxPhotosReached: 'Můžete přidat maximálně 5 fotek',
+            maxPhotosOvershoot: 'Protože můžete přidat maximálně 5 fotek, nebyly všechny vybrané fotky vloženy',
+            noPhotos: 'Zatím jste nepřidali žádné fotky',
+            remove: 'Odstranit',
+        },
+        export: {
+            sending: 'Odesílání',
+            emailIsBeingSend: 'Email se odesílá',
+            emailSuccessfullySent: 'Email byl úspěšně odeslán!',
+            doYouWantToSend: 'Odeslat?',
+            missingField: 'Nevyplněné pole',
+            pleaseFillInField: t`Pro odeslání prosím vyplňte ${'field'}.`,
+            doYouReallyWantToSend: t`Opravdu chcete odeslat email na "${'email'}"?`,
+            emailNotSent: t`Omlouvám se, ale email se nepodařilo odeslat. ${'chyba'}`,
+            doYouRellyWantToRemoveData: 'Chcete odstranit všechna data?',
+            youAreOffline: 'Pravděpodobně nejste připojeni k internetu. Zkontrolujte připojení a zkuste to znovu',
+            errorReported: 'Chyba byla nahlášena.',
+            send: 'Odeslat',
+            nameAndSurnameNeeded: 'Je potřeba zadat alespoň jméno a příjmení',
+            removeAll: 'Odstranit vše',
+            thisIsTheIssue: 'Toto je chyba:',
+            moreInfo: 'Podrobnější informace',
+        },
+        // internetConnectionNeeded: 'Je potřeba připojení k internetu',
+        chooseUnits: 'Vybrat jednotky',
+        remove: 'Odebrat',
+        add: 'Přidat',
+        currency: 'Kč',
+    },
+    units: {
+        kW: 'kW',
+        kWh: 'kWh',
+        m2: 'm²',
+        q: 'q',
+        m3: 'm³',
+        dm3: 'dm³',
+        m: 'm',
+        degreeCelsius: '°C',
+    },
+};
