@@ -1,7 +1,11 @@
-import { template as t } from "$lib/helpers/templates"
-import type { PlainTranslations } from "$lib/translations"
+import { template as t } from '$lib/helpers/templates';
+import type { PlainTranslations } from '$lib/translations';
+import cs from './cs';
+import en from './en';
 
 const de: PlainTranslations = {
+    ...cs,
+    ...en,
     name: `Vorname`,
     surname: `Nachname`,
     phone: `Telefonnummer`,
@@ -9,24 +13,29 @@ const de: PlainTranslations = {
     birthday: `Geburtsdatum`,
     street: `Straße und/oder Haus Nummer`,
     controllerRegistration: `Registrierung eines neuen Reglers`,
-    longAppName: `Verwaltung der IR-Regler`,
-    appName: `Reglerverwaltung IR`,
+    longAppName: `SEIR – Verwaltung der IR-Regler`,
+    appName: `SEIR`,
     requiredField: `Dieses Feld ist erforderlich`,
     controllerType: `Reglertyp`,
     serialNumber: `Seriennummer des Reglers`,
     serialNumberIndoor: `Seriennummer der Inneneinheit`,
     newSerialNumber: `Neue Seriennummer`,
     wrongNumberFormat: `Falsches Zahlenformat`,
-    heatPumpType: `Wärmepumpentyp`,
+    heatPumpType: `Typ der Wärmepumpe`,
+    heatPumpsType: `Typ der Wärmepumpen`,
     airToWater: `Luft/Wasser`,
     groundToWater: `Erdreich/Wasser`,
-    heatPumpModel: `Wärmepumpenmodell`,
-    heatPumpManufactureNumber: `Fertigungsnummer der Wärmepumpe`,
+    heatPumpModelNr: t`${0}Wärmepumpenmodell`,
+    heatPumpManufactureNumberNr: t`Fertigungsnummer der ${0}Wärmepumpe`,
+    first: `erste`,
+    second: `zweite`,
+    third: `dritte`,
+    fourth: `fierte`,
     endUser: `Endbenutzer`,
     wrongDateFormat: `Falsches Datumsformat`,
     wrongPhoneFormat: `Falsches Telefonformat`,
     wrongEmailFormat: `Falsches E-Mail-Format`,
-    roalizationLocation: `Ort der Ausführung`,
+    realizationLocation: `Ort der Ausführung`,
     town: `Stadt`,
     zip: `Postleitzahl`,
     wrongZIPFormat: `Falsches PLZ-Format`,
@@ -57,21 +66,23 @@ const de: PlainTranslations = {
     tooManyRequests: `Zu viele Anfragen! Bitte warten Sie einen Moment`,
     somethingWentWrong: `Etwas ist schiefgelaufen :\\`,
     passwordTooWeak: `Das Passwort ist zu schwach!`,
-    pleaseUseBuisnessEmail: `Bitte verwenden Sie Ihre Firmen-E-Mail-Adresse`,
+    pleaseUseBusinessEmail: `Bitte verwenden Sie Ihre Firmen-E-Mail-Adresse`,
     emailInUse: `Dieses Konto existiert bereits`,
     passwordsDoNotMatch: `Passwörter stimmen nicht überein!`,
     save: `Speichern`,
+    saveAndSend: `Speichern und E-Mails senden`,
     saving: `Speichern...`,
-    youHaveAMistake: `Im Formular liegt ein Fehler vor. Bitte überprüfen Sie die eingegebenen Informationen.`,
+    youHaveAMistake: t`Im Formular liegt ein Fehler vor. Bitte überprüfen Sie die eingegebenen Informationen. Falsche Felder: ${'fields'}`,
     logInNeeded: `Zum Anzeigen und Ausfüllen des Formulars ist eine Anmeldung erforderlich!`,
     evidenceDetailsHtml: t`Details zur Installation mit ${'irType'} ${'irNumber'}`,
     evidenceDetails: `Details zur Installation`,
     loadingData: `Daten werden geladen...`,
     sorrySomethingWentWrong: `Entschuldigung, etwas ist schiefgelaufen.`,
+    somethingWentWrongContactUsHtml: `Entschuldigung, etwas ist schiefgelaufen, bitte kontaktieren Sie uns unter <a href="mailto:Regulus SEIR<aplikace.regulus@gmail.com>?subject=Fehler beim Speichern eines Formulars" target="_blank">aplikace.regulus@gmail.com</a>`,
     linkInvalid: `Entweder ist der Link zu dieser Seite falsch oder der Eintrag wurde bereits gelöscht.`,
     linkToThis: `Link zu dieser Seite`,
     copy: `Kopieren`,
-    regulusRouteForm: `Formular für den Zugriff auf den Regler durch den RegulusRoute-Dienst`,
+    regulusRouteForm: `RegulusRoute Zustimmung`,
     regulusRouteTitle: `Zustimmung zur Bereitstellung des IR-Controllers RegulusRoute`,
     regulusRouteFileName: `RegulusRoute Zustimmung`,
     openPdf: `PDF öffnen`,
@@ -90,6 +101,8 @@ const de: PlainTranslations = {
     residence: `Wohnung`,
     whatToAddInfoTo: `Wozu möchten Sie genauere Informationen hinzufügen?`,
     heatPump: `Wärmepumpe`,
+    heatPumps: `Wärmepumpen`,
+    noPump: `Kein`,
     pleaseFillInIrType: `Bitte geben Sie den IR-Typ ein.`,
     fillInPassword: `Bitte geben Sie das Passwort ein.`,
     search: `Suchen`,
@@ -100,18 +113,18 @@ const de: PlainTranslations = {
     changeWentWrong: `Etwas ist schiefgelaufen. Aktualisieren Sie die Seite und versuchen Sie es erneut.`,
     note: `Hinweis`,
     routeGuide: `Anleitung zur Anmeldung zum Regler`,
-    warranty: `Garantie für die Wärmepumpe`,
-    commissionProtocol: `Inbetriebnahmeprotokoll für die Wärmepumpe`,
-    commissionProtocolTitle: `Protokoll zur Inbetriebnahme der Wärmepumpe `,
-    commissionProtocolFileName: `Wärmepumpe Inbetriebnahmeprotokoll`,
-    installationApproval: `Installationsgenehmigung für die Wärmepumpe`,
-    yearlyCheck: `Jährliche Überprüfung der Wärmepumpe`,
+    warrantyNr: t`Garantie für die Wärmepumpe ${0}`,
+    heatPumpCommissionProtocol: `Inbetriebnahmeprotokoll für die Wärmepumpe`,
+    heatPumpCommissionProtocolTitle: `Protokoll zur Inbetriebnahme der Wärmepumpe `,
+    heatPumpCommissionProtocolFileName: `Wärmepumpe Inbetriebnahmeprotokoll`,
+    yearlyHPCheck: `Jährliche Überprüfung der Wärmepumpe`,
     yearlyCheckFileName: `Jährlichen Inspektion.pdf`,
     yearlyCheckTitle: `Aufgabenbeschreibung bei Durchführung der jährlichen vorbeugenden Inspektion der Luft-Wärmepumpe`,
     doYearlyCheck: `Jährliche Überprüfung der Wärmepumpe durchführen`,
     filledYearlyCheck: `Durchgeführte jährliche Überprüfung der Wärmepumpe`,
-    chooseFromList: `Aus der Liste auswählen:`,
-    chooseCompanyFromList: `Wählen Sie eine Firma aus der Liste aus:`,
+    searchCompanyInList: `Der Firma in der Liste suchen`,
+    searchRepresentative: 'Kontaktperson suchen',
+    representative: 'Kontaktperson',
     newPassword: `Neues Passwort`,
     sendConfirmEmail: `Bestätigungs-E-Mail senden`,
     passwordReset: `Passwort zurücksetzen`,
@@ -141,7 +154,7 @@ const de: PlainTranslations = {
     hpWarranty: `Garantieschreiben für Wärmepumpe`,
     warrantyFileName: `Garantieschreiben.pdf`,
     performingPerson: `Durchführende Person`,
-    realisedAtResidence: `Gleich wie Wohnsitz`,
+    samePlaceAsResidence: `Gleich wie Wohnsitz`,
     agreeWIthRRPrice: `Ich stimme dem einmaligen Preis von 50 € inklusive MwSt. für diese Dienstleistung zu.`,
     suits: `entspricht`,
     suitsNot: `nicht entspricht`,
@@ -193,8 +206,53 @@ const de: PlainTranslations = {
     wasControllerSetToParameters: `Wurde die Regulationseinheit der WP auf die vorgeschriebenen Parameter eingestellt?`,
     wasOwnerFamiliarizedWithFunction: `Wurde der Besitzer in die Grundfunktionen der WP und ihre Bedienung eingewiesen?`,
     isExtendedWarrantyDesired: `Hat der Eigentümer der Wärmepumpe Interesse an einer verlängerten Garantie?`,
-    extendedWarranty7Years: `ja – 7 Jahre auf Kompressor`,
-    extendedWarranty10Years: `ja – 10 Jahre auf Kompressor (gegen Aufpreis)`,
-    isInstallationInWarrantyConditions: `Entsprechen die Installation und Inbetriebnahme den Bedingungen der verlängerten Garantie?`,    
-}
-export default de
+    isInstallationInWarrantyConditions: `Entsprechen die Installation und Inbetriebnahme den Bedingungen der verlängerten Garantie?`,
+    offline: `Sie sind offline!`,
+    mandatoryFields: `Pflichtfelder`,
+    cascadeSee: `Kaskade – siehe unten`,
+    cascade: `Kaskade:`,
+    pumpDetails: t`WP${'n'}: ${'model'} – ${'cislo'}`,
+    check: {
+        kontrolniElektroinstalace: 'Verkabelung der Steuerung',
+        kontrolniUkonyTepelnehoCerpadla: 'Aufgaben zur Inspektion von Wärmepumpen',
+        kontrolniUkonyRegulace: 'Regulierungskontrolle',
+        kontrolniUkonyOtopneSoustavy: 'Inspektionsaufgaben für Heizungsanlagen',
+        kontrolaZasobnikuTv: 'Prüfen des Warmwasserspeichers',
+        kontrolaFunkceVsechElektrickychSpotrebicuZapojenychDoRegulace: 'Überprüfung der Funktion aller an die Steuerung angeschlossenen elektrischen Geräte',
+        kontrolaDotazeniSvorkovychSpoju: 'Anzugskontrolle der Klemmverbindungen',
+        vizualniKontrolaVsechPristupnychVodicuVInstalaciNataveniMechPoskozeni: 'Sichtprüfung aller zugänglichen Drähte in der Anlage (Schmelzen, mechanische Beschädigung)',
+        kontrolaSepnutiDohrevuSepnutiStykacePripadneRele: 'Überprüfung des Einschaltens der Nachheizanlage (Einschalten des Schützes oder des Relais)',
+        kontrolaChoduKompresoru: 'Betrieb des Kompressors prüfen',
+        optickaKontrolaTesnostiTrubkovychSpojuJednotkyAChladivovehoOkruhu: 'Optische Dichtheitsprüfung des Geräts und der Rohrverbindungen des Kältemittelkreislaufs',
+        kontrolaOdvoduKondenzatu: 'Kontrolle des Kondensatablaufs',
+        kontrolaUchyceniVentilatoru: 'Kontrolle der Ventilatorbefestigung',
+        vycisteniVzduchovychCestJednotky: 'Kontrolle der Verdampferflügel, eventuelle Entfernung von Verunreinigungen durch Luft',
+        kontrolaLamelVyparnikuPripadneOdstraneniNecistotVzduchem: 'Kontrolle der Temperaturdifferenz der Wärmepumpe gemäß den Anweisungen',
+        proveritZdaNicNebraniOptimalniCirkulaciVzduchu: 'Kontrolle, dass nichts die optimale Luftzirkulation behindert wird',
+        kontrolaTeplotnihoRozdiluTepelnehoCerpadlaDleNavodu: 'Reinigung der Luftwege des Geräts',
+        kontrolaElektrickeCastiJednotkyTepelnehoCerpadla: 'Überprüfung des elektrischen Teils der Wärmepumpeneinheit',
+        kontrolaChybovychAInformacnichHlaseniRegulatoruAJejichPricin: 'Überprüfung der Fehler- und Informationsmeldungen des Reglers und ihrer Ursachen',
+        kontrolaNastaveniParametruRegulatoru: 'Überprüfung der Parametereinstellungen des Reglers',
+        preventivniProskoleniObsluhyZHlediskaUzivatelskehoNastaveni: 'Vorbeugende Bedienerschulung in Bezug auf die Benutzereinstellungen',
+        stavPocitadlaCelkovychProvoznichHodinKompresoru: 'Zählerstand der Gesamtbetriebsstunden des Verdichters',
+        stavPocitadlaProvoznichHodinDoTvUmoznujeLiToRegulace: 'Status des Betriebsstundenzählers zum WW (wenn die Regelung dies zulässt)',
+        stavCelkovehoPoctuStartuTepCerpadlaUmoznujeLiToRegulace: 'Status der Gesamtzahl der Wärmepumpenstarts (wenn von der Steuerung freigegeben)',
+        stavPoctuStartuTepelCerpadlaDoTvUmoznujeLiToRegulace: 'Status der Anzahl der Wärmepumpenstarts zum WW (wenn von der Regelung freigegeben)',
+        stavPocitadlaCelkovychProvoznichHodinDoplnkovehoZdroje: 'Gesamtbetriebsstundenzählerstand der Zusatzstromversorgung',
+        stavPocitadlaCelkovychProvoznichHodinDoplnkovehoZdrojeTv: 'Status des Gesamtbetriebsstundenzählers der Zusatzstromquelle WW',
+        prumernaCelkovaDobaChoduKompresoruMinOdPosledniKontroly: 'Durchschnittliche Gesamtlaufzeit des Kompressors [min] - seit der letzten Kontrolle',
+        prumernaDobaChoduKompresoruDoTvMinOdPosledniKontroly: 'Durchschnittliche Kompressorlaufzeit zu WW [min] - seit der letzten Kontrolle',
+        kontrolaFunkceObehovychCerpadel: 'Kontrolle der Funktion von Umwälzpumpen',
+        vycisteniFiltruObehovychCerpadel: 'Reinigung der Umwälzfilter',
+        odvzdusneniZdrojeTc: 'Entlüftung der Quelle (Wärmepumpe)',
+        kontrolaFunkceVsechMotorickychVentiluSmesovaciZonovych: 'Funktionskontrolle aller motorisierten Ventile (Misch- und Zonenventile)',
+        kontrolaTesnostiOtopneSoustavy: 'Überprüfung der Heizungsanlage auf Dichtheit',
+        kontrolaTlakuVExpanzniNadobeOtopneSoustavy: 'Kontrolle des Drucks im Ausdehnungsgefäß der Heizungsanlage',
+        nastavenyTlakPriUvadeniDoProvozu: 'Druck bei der Inbetriebnahme einstellen',
+        pripadneProvedteKontroluTlakuVOtopneSoustave: 'Alternativ kann der Druck im Heizungssystem überprüft werden',
+        kontrolaMgAnodyVZasobnikuPripVymena: 'Kontrolle der Magnesiumanode im Behälter, ggf. Austausch',
+        kontrolaPojistovacihoVentilu: 'Kontrolle des Sicherheitsventils',
+        pripadneProvedteKontroluTlakuVEnTepleVody: 'Warmwasserdruck prüfen, falls erforderlich',
+    },
+};
+export default de;
