@@ -1,21 +1,29 @@
 import type { PlainTranslations } from "$lib/translations"
 import { template as t } from "$lib/helpers/templates"
+import cs from "./cs"
 
 const en: PlainTranslations = {
+    ...cs,
     controllerRegistration: `New controller registration`,
-    longAppName: "IR Controller Registration",
-    appName: "IR Registration",
+    longAppName: "SEIR – IR Controller Registration",
+    appName: "SEIR",
     requiredField: 'This field is required',
     controllerType: 'Controller type',
     serialNumber: 'Controller serial number',
     serialNumberIndoor: 'Indoor unit serial number',
     newSerialNumber: 'New serial number',
     wrongNumberFormat: 'Invalid number format',
+    noPump: 'None',
     heatPumpType: 'Heat pump type',
+    heatPumpsType: 'Type of heat pumps',
     airToWater: 'Air-to-Water',
     groundToWater: 'Ground-to-Water',
-    heatPumpModel: 'Heat pump model',
-    heatPumpManufactureNumber: 'Heat pump manufacture number',
+    heatPumpModelNr: t`${0}heat pump model`,
+    heatPumpManufactureNumberNr: t`${0}heat pump manufacture number`,
+    first: `first`,
+    second: `second`,
+    third: `third`,
+    fourth: `fourth`,
     endUser: 'End user',
     name: 'Name',
     surname: 'Surname',
@@ -25,7 +33,7 @@ const en: PlainTranslations = {
     wrongPhoneFormat: 'Invalid phone number format',
     email: 'Email',
     wrongEmailFormat: 'Invalid email format',
-    roalizationLocation: 'Realization location',
+    realizationLocation: 'Realization location',
     town: 'Municipality',
     street: 'Street name and/or building number',
     zip: 'Postal (ZIP) code',
@@ -57,23 +65,25 @@ const en: PlainTranslations = {
     tooManyRequests: 'Too many requests! Please wait a bit',
     somethingWentWrong: 'Something went wrong :\\',
     passwordTooWeak: 'Password is too weak!',
-    pleaseUseBuisnessEmail: 'Please use your buisness email.',
+    pleaseUseBusinessEmail: 'Please use your business email.',
     emailInUse: 'This account already exists',
     passwordsDoNotMatch: "Passwords don't match!",
     save: 'Save',
+    saveAndSend: 'Save & send emails',
     saving: 'Saving...',
-    youHaveAMistake: 'You have a mistake in the form. Please, check the filled information.',
+    youHaveAMistake: t`You have a mistake in the form. Please, check the filled information. Invalid fields: ${'fields'}`,
     logInNeeded: 'You must log in to view and fill out the form!',
     evidenceDetailsHtml: t`Installation with ${'irType'} ${'irNumber'} details`,
     evidenceDetails: `Installation details`,
     loadingData: 'Loading data...',
     sorrySomethingWentWrong: 'Apologize, something went wrong.',
+    somethingWentWrongContactUsHtml: `Apologize, something went wrong, please contact us at <a href="mailto:Regulus SEIR<aplikace.regulus@gmail.com>?subject=Error saving a form" target="_blank">aplikace.regulus@gmail.com</a>`,
     linkInvalid: 'Either the link to this page is incorrect or the registration record has already been removed.',
     linkToThis: 'Lint to this page',
     copy: 'Copy',
-    regulusRouteForm: `Form on making the controller available to the RegulusRoute service`,
+    regulusRouteForm: `Approval with RegulusRoute`,
     regulusRouteTitle: `Consent to making the IR controller available to the RegulusRoute service`,
-    regulusRouteFileName: `RegulusRoute form.pdf`,
+    regulusRouteFileName: `Approval with RegulusRoute.pdf`,
     openPdf: 'Open pdf',
     changeLang: 'Change language',
     deleteThisEvidence: 'Remove this registration record',
@@ -90,6 +100,7 @@ const en: PlainTranslations = {
     residence: 'Place of residence',
     whatToAddInfoTo: 'What would you like to add more detailed information to?',
     heatPump: 'Heat pump',
+    heatPumps: 'Heat pumps',
     pleaseFillInIrType: 'Please, fill in the IR type',
     fillInPassword: 'Please, enter your password.',
     search: 'Search',
@@ -100,16 +111,16 @@ const en: PlainTranslations = {
     changeWentWrong: 'Something went wrong. Refresh the page and try again.',
     note: 'Note',
     routeGuide: 'Instructions for logging into the controller',
-    warranty: 'Heat pump warranty',
-    commissionProtocol: 'Heat pump instalation protocol',
-    installationApproval: 'Heat pump instalation approval',
-    yearlyCheck: 'Yearly heat pump check',
+    warrantyNr: t`Heat pump ${0} warranty`,
+    heatPumpCommissionProtocol: 'Heat pump instalation protocol',
+    yearlyHPCheck: 'Yearly heat pump check',
     doYearlyCheck: 'Do a yearly heat pump check',
     filledYearlyCheck: 'Yearly heat pump checks',
     yearlyCheckFileName: `Yearly check.pdf`,
     yearlyCheckTitle: `Description of operations during the preventive annual inspection of an air heat pump`,
-    chooseFromList: 'Choose from list',
-    chooseCompanyFromList: 'Choose company from list:',
+    searchCompanyInList: 'Search company in list',
+    searchRepresentative: 'Search representative',
+    representative: 'Representative',
     newPassword: 'New Password',
     sendConfirmEmail: 'Send confirmation email',
     emailNotSent: t`The email could not be sent: ${'status'} ${'statusText'}`,
@@ -139,6 +150,14 @@ const en: PlainTranslations = {
     hpWarranty: `Heat pump warranty`,
     warrantyFileName: `Warranty.pdf`,
     performingPerson: `Performing person`,
-    realisedAtResidence: 'Same as the place of residence',
+    samePlaceAsResidence: 'Same as the place of residence',
+    offline: `You are offline!`,
+    successfullyDeleted: `Record successfully deleted.`,
+    mandatoryFields: `mandatory fields`,
+    theme: {
+        dark: 'Dark',
+        light: 'Light',
+        auto: 'Auto',
+    },
 }
 export default en
