@@ -1,5 +1,223 @@
-import type { PlainTranslations } from "$lib/translations"
-import cs from "./cs"
+import type { PlainTranslations } from '$lib/translations';
+import cs from './cs';
+import { template as t } from '$lib/helpers/templates';
 
-const sk: PlainTranslations = cs
-export default sk
+const sk: PlainTranslations = {
+    ...cs,
+    note: `Poznámka`,
+    yes: `Áno`,
+    no: `Nie`,
+    cancel: `Zrušiť`,
+    demand: {
+        ...cs.demand,
+        contacts: {
+            contacts: `Kontakty a miesto realizácie`,
+            surname: `Priezvisko`,
+            name: `Meno`,
+            street: `Ulica, č. p.`,
+            city: `Mesto`,
+            zip: `PSČ`,
+            phone: `Telefón`,
+            email: `Email`,
+            chooseAssemblyCompanyHere: `Vyberte montážnu firmu:`,
+            chooseCompany: `IČO monážní firmy`,
+            chooseAssemblyCompany: `Vybrat firmu`,
+            searchCompany: `Vyberte montážnu firmu`,
+            crn: `Vyhledat firmu`,
+            demandOrigin: `Pôvod dopytu`,
+            originQuestionEmail: `Otázka poslaná emailom od koncového zákazníka`,
+            originQuestionExhibition: `Otázka z výstavy`,
+            originQuestionInPerson: `Otázka pri osobnej návšteve u koncového zákazníka`,
+            originDistributionCompany: `Dopyt od distribučnej firmy`,
+            originAssembleres: `Dopyt od montážnikov`,
+            originDesigner: `Dopyt od projektanta`,
+            demandSubject: `Predmet dopytu`,
+            heatPump: `Tepelné čerpadlo`,
+            fve: `FVE`,
+        },
+        objectDetail: {
+            objectDetail: `Detail objektu`,
+            heatLoss: `Tepelná strata`,
+            heatNeedsForHeating: `Potreba tepla – kúrenie`,
+            heatNeedsForHotWater: `Potreba tepla – teplá voda`,
+            area: `Vykurovaná plocha`,
+            volume: `Vykurovaný objem`,
+            costs: `Náklady na kúrenie`,
+            type: `Druh paliva`,
+            usage: `Spotreba paliva`,
+            type2: `Druh 2. paliva`,
+            usage2: `Spotreba 2. paliva`,
+        },
+        fve: {
+            fve: `FVE`,
+            currentSituation: `Súčasný stav`,
+            currentHeating: `Súčasný zdroj vykurovania`,
+            currentHotWater: `Súčasný zdroj prípravy teplej vody`,
+            currentTanks: `Inštalované nádrže/zásobníky`,
+            currentConsumption: `Aktuálna ročná spotreba el. energie`,
+            breakerSize: `Veľkosť hlavného ističa`,
+            tariff: `Sadzba`,
+            breakerBoxLocation: `Umiestnenie domového rozvádzača`,
+            requirements: `Požiadavky na FVE`,
+            requiredPower: `Požadovaný výkon`,
+            locationBuidingType: `Typ budovy na inštaláciu kolektorov`,
+            lightningRod: `Na streche pre inštaláciu je umiestnený bleskozvod`,
+            familyHouseEtc: `Rodinný dom, stodola, pergola, …`,
+            avaiableAreas: `Dostupné plochy pre inštaláciu`,
+            roofMaterial: `Materiál stresnej krytiny`,
+            roofAge: `Vek strešnej krytiny`,
+            tile: `Taška`,
+            tileType: `Typ tašiek`,
+            metalSheetFolded: `Plech falcovaný`,
+            metalSheetTrapezoidal: `Plech trapézový`,
+            foil: `Fólia`,
+            asphaltShingle: `Asfaltový šindeľ`,
+            useOptimizers: `Použiť optimizéry`,
+            size: t`Rozmer ${0}. plochy`,
+            orientation: t`Orientácia ${0}. plochy`,
+            slope: t`Sklon ${0}. plochy`,
+            battery: `Akumulácia do batérií`,
+            batteryCapacity: `Kapacita batérií`,
+            water: `Akumulácia do vody`,
+            network: `Pretoky do siete`,
+            networkPower: `Pretoky do siete – rezervovaný výkon`,
+            charging: `Dobíjacia stanica – wallbox`,
+        },
+        system: {
+            system: `Systém`,
+            hPType: `Typ tepelného čerpadla`,
+            hPModel: `Model tepelného čerpadla`,
+            storeType: `Typ nádrže`,
+            storeNone: 'Žiadná',
+            storeVolume: `Objem nádrže`,
+            tankType: `Typ zásobníka`,
+            tankNone: 'Žiadná',
+            tankVolume: `Objem zásobníka`,
+            indoorUnitType: `Typ vnútornej jednotky`,
+            indoorUnitNone: 'Žiadná',
+            heatingSystem: `vykurovací systém`,
+            heatingSystem1circuit: `1 vykurovací okruh`,
+            heatingSystem2circuits: `2 vykurovacie okruhy`,
+            heatingSystem3circuits: `3 vykurovacie okruhy`,
+            heatingSystemInvertor: `Invertor na priamo`,
+            heatingSystemOther: 'Iný',
+            hotWaterCirculation: `Cirkulácia teplej vody`,
+            poolHeating: `Ohrev bazéna`,
+            airToWater: `vzduch/voda`,
+            groundToWater: `zem/voda`,
+        },
+        pool: {
+            pool: `Bazén`,
+            usagePeriod: `Doba využívania`,
+            periodYearlong: `celoročne`,
+            periodSeasonal: `sezónne`,
+            location: `Umiestnenie`,
+            locationOutdoor: `vonkajší`,
+            locationIndoor: `vnútorný`,
+            waterType: `Druh vody`,
+            freshType: `sladká`,
+            saltType: `slaná`,
+            shape: `tvar`,
+            shapeRectangle: `obdĺžnikový`,
+            shapeOval: `oválný`,
+            shapeCircle: `kruhový`,
+            length: `Dĺžka`,
+            width: `Šírka`,
+            radius: `Priemer`,
+            depth: `Hĺbka`,
+            coverage: `Zakrytie`,
+            coverageNone: 'žiadné',
+            coverageSolid: `pevná strecha`,
+            coveragePolycarbonate: `polykarbonát`,
+            coverageOther: 'iné',
+            temperature: `požadovaná teplota`,
+        },
+        additionalSources: {
+            sources: `Doplnkové zdroje`,
+            hotWater: `Teplá voda`,
+            heating: `Kúrenie`,
+            heatingElement: `Vykurovacie teleso v nádrži`,
+            electricBoiler: `Elektrokotol`,
+            gasBoiler: `Plynový kotol`,
+            fireplace: `Krb alebo kotol na tuhé paliva`,
+            newNeuter: `nové`,
+            newMasculine: `nový`,
+            existing: `existujúce`,
+            toSocket: `do zásuvky`,
+            fromRegulation: `ovládané z regulácie`,
+            otherSource: 'Iný',
+        },
+        accessories: {
+            accessories: `Příslušenství`,
+            hose: `Hadice`,
+            wallSupportBracket: `Držák pro TČ`,
+            roomUnitsAndSensors: `Pokojové jednotky a čidla`,
+            heatingCable: `Topný kabel`,
+            onWall: `Na stěnu`,
+            onIsolatedWall: `Na izolovanou stěnu`,
+        },
+        logIn: {
+            logIn: `Prihlásenie`,
+            internetNeeded: `Pre prihlásenie je potrebné pripojenie k internetu!`,
+            iAmEmploee: `Som zamestanec Regulusu`,
+            iAmNotEmploee: `Nie som zamestanec Regulusu`,
+            yourRepresentative: `Váš obchodný zástupca`,
+            yourName: `Vaše meno`,
+            yourSurname: `Vaše priezvisko`,
+            yourCrn: `Vaše IČO (nepovinné)`,
+            yourEmail: `Váš email`,
+            chooseYourself: `Vyber sa:`,
+            selectedFullName: t`Meno a priezvisko: ${'jmeno'} ${'prijmeni'}`,
+            selectedEmail: t`Email: ${'email'}`,
+            selectedCode: t`Číslo KO: ${'kod'}`,
+            selectedCrn: t`IČO: ${'ico'}`,
+            nameAndSurname: `Meno a priezvisko:`,
+            email: `Email:`,
+            code: `Číslo KO:`,
+            crn: `IČO:`,
+            addMoreInfo: `Doplňte, prosím, ešte nejaké informácie o Vás:`,
+            representativeNeeded: `Je potrebné vybrať obchodného zástupcu`,
+            emailNeeded: `Je potrebné vyplniť email`,
+            surnameNeeded: `Je potrebné vyplniť priezvisko`,
+            nameNeeded: `Je potrebné vyplniť meno`,
+            youAreNotLoggedIn: `Nie ste prihlásený!`,
+            logOut: `Odhlásiť sa`,
+        },
+        photos: {
+            photosManager: `Správa fotiek`,
+            takePhoto: `Odfotiť`,
+            choosePhoto: `Vybrať z galérie`,
+            photo: `Fotka`,
+            add: `Pridať fotku`,
+            maxPhotosReached: `Môžete pridať maximálne 5 fotiek`,
+            maxPhotosOvershoot: `Pretože môžete pridať maximálne 5 fotiek, neboli všetky vybrané fotky vložené`,
+            noPhotos: `Zatiaľ ste nepridali žiadne fotky`,
+            remove: `Odstrániť`,
+        },
+        export: {
+            sending: `Odosielanie`,
+            missingField: `Nevyplnené pole`,
+            pleaseFillInField: t`Pre odoslanie prosím vyplňte ${'field'}.`,
+            emailIsBeingSend: `Email sa odosiela`,
+            emailSuccessfullySent: `Email bol úspešne odoslaný!`,
+            doYouWantToSend: `Odoslať?`,
+            doYouReallyWantToSend: t`Naozaj chcete odoslať email na \`${'email'}\`?`,
+            emailNotSent: t`Ospravedlňujem sa, ale email sa nepodarilo odoslať. ${'chyba'}`,
+            doYouRellyWantToRemoveData: `Chcete odstrániť všetky dáta?`,
+            youAreOffline: `Pravdepodobne nie ste pripojení k internetu. Skontrolujte pripojenie a skúste to znovu`,
+            errorReported: `Chyba bola nahlásená.`,
+            send: `Odoslať`,
+            nameAndSurnameNeeded: `Je potrebné zadať aspoň meno a priezvisko`,
+            removeAll: `Odstrániť všetko`,
+            thisIsTheIssue: `Toto je chyba:`,
+            moreInfo: `Podrobnejšie informácie`,
+        },
+        // internetConnectionNeeded: `Je potrebné pripojenie k internetu`,
+        chooseUnits: `Vybrať jednotky`,
+        remove: `Odobrať`,
+        add: `Pridať`,
+        currency: '€',
+    },
+};
+
+export default sk;
