@@ -1,6 +1,8 @@
-import { checkAuth } from "$lib/client/auth";
+import { checkAuth, currentUser } from '$lib/client/auth';
 import type { LayoutLoad } from "./$types";
+import { get } from 'svelte/store';
 
 export const load: LayoutLoad = async () => ({
-    isLoggedIn: await checkAuth()
-})
+    isLoggedIn: await checkAuth(),
+    user: get(currentUser),
+} as const)
