@@ -1,5 +1,5 @@
 import type { PageLoad } from './$types';
-import { checkAdmin, checkAuth } from '$lib/client/auth';
+import { checkAdmin, checkAuth, checkRegulusOrAdmin } from '$lib/client/auth';
 import { languageCodes } from "$lib/languages";
 import type { EntryGenerator } from "./$types";
 import { browser } from '$app/environment';
@@ -11,5 +11,6 @@ export const entries: EntryGenerator = () => [
 ];
 
 export const load: PageLoad = async () => {
+	console.log(await checkRegulusOrAdmin())
     if ((!await checkAuth() || !await checkAdmin()) && browser) error(401)
 }
