@@ -1,23 +1,30 @@
 import adapter from '@sveltejs/adapter-node';
 import { sveltePreprocess } from 'svelte-preprocess';
+import { execSync } from 'node:child_process';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	preprocess: sveltePreprocess({
 		preserve: ['module']
 	}),
+	onwarn: (warning, handler) => {
+		if (warning.code === 'css-unused-selector') {
+			return;
+		}
+		handler(warning);
+	},
 	kit: {
 		adapter: adapter({
 			out: 'dist'
 		}),
-		// files: {
-		// 	serviceWorker: 'src/service-worker.ts'
-		// },
-		
+		version: {
+			name: execSync('git rev-parse HEAD').toString().trim()
+		},
+		csrf: {
+			checkOrigin: false,
+		}
 	},
-	compilerOptions: {
-		
-	},
+	compilerOptions: {},
 };
 
 export default config;
